채플 현재 학기 반영, 채플 DB 구조 수정, rusaintAPI 버전업 채플 관련 텍스트 추가, 예외 처리 추가

+++ app/src/main/kotlin/com/yourssu/soomsil/usaint/core/model/ChapelAttendanceData.kt
@@ -1,7 +1,7 @@
 package com.yourssu.soomsil.usaint.core.model
 
 data class ChapelAttendanceData(
-    val division: Int,
+    val division: Long,
     val classDate: String,
     val category: String,
     val instructor: String,
